pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "frustrampnn"
version = "1.0.0"
description = "Ultra-fast deep learning prediction of single-residue local energetic frustration"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "Max Beining" },
    { name = "Felipe Engelberger", email = "felipeengelberger@gmail.com" },
    { name = "Clara T. Schoeder" },
    { name = "César A. Ramírez-Sarmiento" },
    { name = "Jens Meiler" },
]
keywords = [
    "protein",
    "frustration",
    "deep-learning",
    "bioinformatics",
    "structural-biology",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
]
dependencies = [
    "torch>=2.0.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "omegaconf>=2.3.0",
    "tqdm>=4.65.0",
    "biopython>=1.80",
    "click>=8.0.0",
    # frustrapy and its dependencies for validation
    "frustrapy @ git+https://github.com/engelberger/frustrapy.git@dev_benchmark",
    "igraph>=1.0.0",
    "leidenalg>=0.10.0",
]

[project.optional-dependencies]
train = [
    "pytorch-lightning>=2.0.0",
    "torchmetrics>=1.0.0",
    "wandb>=0.15.0",
]
esm = [
    "fair-esm==2.0.0",
]
viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "plotly>=5.18.0",
    "kaleido>=0.2.1",
    "py3Dmol>=2.0.0",
]
validation = [
    "scipy>=1.10.0",
]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.0",
    "ruff>=0.5.0",
    "mypy>=1.0",
]
all = [
    "frustrampnn[train,esm,viz,dev]",
]

[project.scripts]
frustrampnn = "frustrampnn.cli:main"

[project.urls]
Homepage = "https://github.com/schoederlab/frustraMPNN"
Documentation = "https://github.com/schoederlab/frustraMPNN#readme"
Repository = "https://github.com/schoederlab/frustraMPNN"
Issues = "https://github.com/schoederlab/frustraMPNN/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/frustrampnn"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "C408",   # unnecessary dict() call - common in plotly code
]

[tool.ruff.lint.isort]
known-first-party = ["frustrampnn"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --tb=short"

